backend/flexiconvert-backend/requirements.txt
Pillow>=10.0.0
pillow-heif>=0.15.0
img2pdf>=0.5.1
cairosvg>=2.7.1 